Factor out common test logic
